disable subset_conus test on travis

+++ tests/subset_conus_tests.py
@@ -90,35 +90,38 @@ class SubsetConusRegressionTests(unittest.TestCase):
 
 
     def test_conus1_subset_regression(self):
-        test_dir = Path('test_outputs')
-        test_dir.mkdir(exist_ok=True)
-        subset_conus.subset_conus(input_path=self.good_shape_file_path,
-                                  shapefile=self.good_shape_file_name, conus_files='/home/arezaii/git/subset_1/CONUS1_inputs',
-                                  out_dir=test_dir,
-                                  out_name='test_conus1_subset')
-        ref_subsurface = read_file(test_files.huc10190004.get('conus1_subsurface'))
-        ref_mask = read_file(test_files.huc10190004.get('conus1_mask'))
-        ref_slopex = read_file(test_files.huc10190004.get('conus1_slopex'))
-        ref_slopey = read_file(test_files.huc10190004.get('conus1_slopey'))
+        if os.environ.get('TRAVIS'):
+            pass
+        else:
+            test_dir = Path('test_outputs')
+            test_dir.mkdir(exist_ok=True)
+            subset_conus.subset_conus(input_path=self.good_shape_file_path,
+                                      shapefile=self.good_shape_file_name, conus_files='/home/arezaii/git/subset_1/CONUS1_inputs',
+                                      out_dir=test_dir,
+                                      out_name='test_conus1_subset')
+            ref_subsurface = read_file(test_files.huc10190004.get('conus1_subsurface'))
+            ref_mask = read_file(test_files.huc10190004.get('conus1_mask'))
+            ref_slopex = read_file(test_files.huc10190004.get('conus1_slopex'))
+            ref_slopey = read_file(test_files.huc10190004.get('conus1_slopey'))
 
-        self.assertIsNone(np.testing.assert_array_equal(ref_subsurface, read_file(test_dir / 'grid3d.v3_clip.pfb')))
-        self.assertIsNone(np.testing.assert_array_equal(ref_mask, read_file(test_dir / 'test_conus1_subset_raster_from_shapefile.tif')))
-        self.assertIsNone(np.testing.assert_array_equal(ref_slopex, read_file(test_dir /'slopex_clip.pfb')))
-        self.assertIsNone(np.testing.assert_array_equal(ref_slopey, read_file(test_dir /'slopey_clip.pfb')))
+            self.assertIsNone(np.testing.assert_array_equal(ref_subsurface, read_file(test_dir / 'grid3d.v3_clip.pfb')))
+            self.assertIsNone(np.testing.assert_array_equal(ref_mask, read_file(test_dir / 'test_conus1_subset_raster_from_shapefile.tif')))
+            self.assertIsNone(np.testing.assert_array_equal(ref_slopex, read_file(test_dir /'slopex_clip.pfb')))
+            self.assertIsNone(np.testing.assert_array_equal(ref_slopey, read_file(test_dir /'slopey_clip.pfb')))
 
-        with open(test_dir / 'test_conus1_subset.pfsol', 'r') as test_file:
-            with open(test_files.huc10190004.get('conus1_sol'), 'r') as ref_file:
-                self.assertEqual(test_file.read(), ref_file.read(),
-                                 'Writing PFSOL file matches reference for conus1')
-        """
-        make sure the vtk files match as well. skip the first two lines as they contain a version and filename that
-        may vary
-        """
-        with open(test_dir / 'test_conus1_subset.vtk', 'r') as test_file:
-            with open(test_files.huc10190004.get('conus1_vtk'), 'r') as ref_file:
-                self.assertEqual(test_file.read().split('\n')[2:], ref_file.read().split('\n')[2:],
-                                 'Writing vtk file matches reference for conus1')
-        shutil.rmtree(test_dir)
+            with open(test_dir / 'test_conus1_subset.pfsol', 'r') as test_file:
+                with open(test_files.huc10190004.get('conus1_sol'), 'r') as ref_file:
+                    self.assertEqual(test_file.read(), ref_file.read(),
+                                     'Writing PFSOL file matches reference for conus1')
+            """
+            make sure the vtk files match as well. skip the first two lines as they contain a version and filename that
+            may vary
+            """
+            with open(test_dir / 'test_conus1_subset.vtk', 'r') as test_file:
+                with open(test_files.huc10190004.get('conus1_vtk'), 'r') as ref_file:
+                    self.assertEqual(test_file.read().split('\n')[2:], ref_file.read().split('\n')[2:],
+                                     'Writing vtk file matches reference for conus1')
+            shutil.rmtree(test_dir)
 
 
 if __name__ == '__main__':
